Moved BaseVisitor
